tests: Improve companies invitation tests
